Complete function game
game() is the main function of the game.
To finish this function, previous function playRound has been reformatted to return an integer that indicates the outcome.
Two helper functions, getRoundOutcomeMessage and getGameOutcomeMessage have been created to generate output messages depending on the outcome of each round and the overall outcome of the game respectively.